Update README.rst based on Eric's feedback
My feedback is a section on "Expectations and Evaluation Criteria" should be added - similar to what in the Autocomplete, but updated with the three things in the rubric for this one.  The things listed under criteria for success are more like Requirements, so could be kept but renamed.  It is important for me to convey up front, how we'll evaluate.